Kubernetes cost optimisation - How to actually do it
Measuring and controlling costs in Kubernetes is often complex. But it does not need to be.

In this practical session with demos, I will show you different ways how you control your cloud bill in Kubernetes. From easy implementations, to policy changes and using products and AI-driven approaches, this session will be your go-to for Kubernetes cost optimisation.